test: Ticket 로또 번호 묶음 테스트 보충, TestProvider 클래스 생성
