+++ src/test/java/lotto/domain/ticket/TicketTest.java
@@ -3,14 +3,11 @@ package lotto.domain.ticket;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.NullSource;
 
@@ -20,7 +17,9 @@ import lotto.exception.ticket.TicketNumbersExceptionStatus;
 
 class TicketTest {
 
-	@DisplayName("숫자 요소는 NULL이 아니어야 합니다.")
+	private static final String PROVIDER_PATH = "lotto.domain.ticket.provider.TicketTestProvider#";
+
+	@DisplayName("로또 번호 묶음은 NULL이 아니어야 합니다.")
 	@ParameterizedTest
 	@NullSource
 	void ticketNullExceptionTest(final List<Integer> numbers) {
@@ -29,43 +28,27 @@ class TicketTest {
 				.hasMessageContaining(TicketNumbersExceptionStatus.NUMBERS_IS_NULL.getMessage());
 	}
 
-	@DisplayName("숫자 요소는 6개여야 합니다.")
-	@ParameterizedTest
-	@MethodSource("provideForNumbersOutOfSizeExceptionTest")
+	@DisplayName("로또 번호 묶음은 6개로 구성되여야 합니다.")
+	@ParameterizedTest(name = "[{index}] 로또 번호 : {0}")
+	@MethodSource(PROVIDER_PATH + "provideForNumbersOutOfSizeExceptionTest")
 	void ticketOutOfSizeExceptionTest(final List<Integer> numbers) {
 		assertThatThrownBy(() -> new Ticket(numbers))
 				.isInstanceOf(LottoException.class)
 				.hasMessageContaining(TicketNumbersExceptionStatus.NUMBERS_OUT_OF_SIZE.getMessage());
 	}
 
-	public static Stream<Arguments> provideForNumbersOutOfSizeExceptionTest() {
-		return Stream.of(
-				Arguments.of(Arrays.asList(1, 2, 3, 4, 5)),
-				Arguments.of(Arrays.asList(1, 2, 3, 4, 5, 6, 7)),
-				Arguments.of(Arrays.asList(1, 2, 3, 4, 5, 6, 7, 8))
-		);
-	}
-
-	@DisplayName("숫자 요소는 중복될 수 없습니다.")
-	@ParameterizedTest
-	@MethodSource("provideForNumbersDuplicatedExceptionTest")
+	@DisplayName("로또 번호 묶음 중 중복 값은 존재할 수 없습니다.")
+	@ParameterizedTest(name = "[{index}] 로또 번호 : {0}")
+	@MethodSource(PROVIDER_PATH + "provideForNumbersDuplicatedExceptionTest")
 	void ticketNumbersDuplicatedExceptionTest(final List<Integer> numbers) {
 		assertThatThrownBy(() -> new Ticket(numbers))
 				.isInstanceOf(LottoException.class)
 				.hasMessageContaining(TicketNumbersExceptionStatus.NUMBERS_DUPLICATED.getMessage());
 	}
 
-	public static Stream<Arguments> provideForNumbersDuplicatedExceptionTest() {
-		return Stream.of(
-				Arguments.of(Arrays.asList(1, 1, 3, 4, 5, 6)),
-				Arguments.of(Arrays.asList(1, 1, 1, 4, 5, 3)),
-				Arguments.of(Arrays.asList(5, 2, 3, 4, 5, 5))
-		);
-	}
-
-	@DisplayName("볼 포함 여부 확인 테스트")
-	@ParameterizedTest
-	@MethodSource("provideForContainsTest")
+	@DisplayName("특정 번호 포함 여부 확인 테스트")
+	@ParameterizedTest(name = "[{index}] 로또 번호 : {0}, 특정 번호 : {1}")
+	@MethodSource(PROVIDER_PATH + "provideForContainsTest")
 	void containsTest(final List<Integer> numbers, final int targetNumber) {
 		final Ticket ticket = new Ticket(numbers);
 		final Ball targetBall = new Ball(targetNumber);
@@ -73,17 +56,9 @@ class TicketTest {
 		assertThat(ticket.contains(targetBall)).isTrue();
 	}
 
-	public static Stream<Arguments> provideForContainsTest() {
-		return Stream.of(
-				Arguments.of(Arrays.asList(1, 2, 3, 4, 5, 6), 1),
-				Arguments.of(Arrays.asList(1, 2, 3, 4, 5, 6), 3),
-				Arguments.of(Arrays.asList(1, 2, 3, 4, 5, 6), 6)
-		);
-	}
-
-	@DisplayName("당첨 번호와 일치 개수 확인")
-	@ParameterizedTest
-	@MethodSource("provideForCountMatchesTest")
+	@DisplayName("당첨 번호와 일치 개수 확인 테스트")
+	@ParameterizedTest(name = "[{index}] 로또 번호 : {0}, 일치 개수 : {2}, 당첨 번호 : {1}")
+	@MethodSource(PROVIDER_PATH + "provideForCountMatchesTest")
 	void countMatches(final List<Integer> numbers, final List<Integer> winningNumbers, final int matchCount) {
 		final Ticket ticket = new Ticket((numbers));
 		final Ticket winningTicket = new Ticket(winningNumbers);
@@ -95,29 +70,9 @@ class TicketTest {
 		assertThat(ticket.countMatches(winningBalls)).isEqualTo(matchCount);
 	}
 
-	public static Stream<Arguments> provideForCountMatchesTest() {
-		return Stream.of(
-				Arguments.of(
-						Arrays.asList(1, 2, 3, 4, 5, 6),
-						Arrays.asList(11, 12, 13, 14, 15, 16),
-						0
-				),
-				Arguments.of(
-						Arrays.asList(1, 2, 3, 4, 5, 6),
-						Arrays.asList(1, 2, 3, 14, 15, 16),
-						3
-				),
-				Arguments.of(
-						Arrays.asList(1, 2, 3, 4, 5, 6),
-						Arrays.asList(1, 2, 3, 4, 5, 6),
-						6
-				)
-		);
-	}
-
 	@DisplayName("당첨 등수 확인 테스트")
-	@ParameterizedTest
-	@MethodSource("provideForCalculateRankTest")
+	@ParameterizedTest(name = "[{index}] 로또 번호 : {0}, 당첨 번호 : {1}, 보너스 번호 : {2}, 당첨 등수 : {3}")
+	@MethodSource(PROVIDER_PATH + "provideForCalculateRankTest")
 	void calculateRankTest(final List<Integer> numbers,
 					 final List<Integer> winningNumbers,
 					 final int bonusNumber,
@@ -127,31 +82,6 @@ class TicketTest {
 		final Ball bonusBall = new Ball(bonusNumber);
 		final Rank rank = ticket.calculateRank(winningTicket, bonusBall).orElse(null);
 		assertThat(rank).isEqualTo(expected);
-	}
-
-	public static Stream<Arguments> provideForCalculateRankTest() {
-		return Stream.of(
-				Arguments.of(
-						Arrays.asList(1, 2, 3, 4, 5, 6),
-						Arrays.asList(11, 12, 13, 14, 15, 16),
-						17, null
-				),
-				Arguments.of(
-						Arrays.asList(1, 2, 3, 4, 5, 6),
-						Arrays.asList(1, 2, 3, 14, 15, 16),
-						17, Rank.FIFTH_GRADE
-				),
-				Arguments.of(
-						Arrays.asList(1, 2, 3, 4, 5, 6),
-						Arrays.asList(1, 2, 3, 4, 5, 16),
-						6, Rank.SECOND_GRADE
-				),
-				Arguments.of(
-						Arrays.asList(1, 2, 3, 4, 5, 6),
-						Arrays.asList(1, 2, 3, 4, 5, 6),
-						7, Rank.FIRST_GRADE
-				)
-		);
 	}
 
 }
